Refactoring checkUser()->isUserLoggedIn(), trying to fix remember_me issue but it doesn't work yet

+++ src/Controller/PinController.php
@@ -23,7 +23,9 @@ class PinController extends AbstractController
      */
     public function index(PinRepository $pinRepo, Request $req): Response
     {
-        $this->checkUser();
+        if(!$this->isUserLoggedIn()){
+            return $this->redirectToRoute('app_login');
+        }
         $pins = $pinRepo->findBy([], ["createdAt" => "DESC"]);
         return $this->render('pin/index.html.twig', compact('pins'));
     }
@@ -33,7 +35,9 @@ class PinController extends AbstractController
      */
     public function create(Request $req, EntityManagerInterface $em): Response
     {
-        $this->checkUser();
+        if(!$this->isUserLoggedIn()){
+            return $this->redirectToRoute('app_login');
+        }
         $pin = new Pin;
         $form = $this->createForm(PinType::class, $pin);
         $form->handleRequest($req);
@@ -53,7 +57,9 @@ class PinController extends AbstractController
      */
     public function update(Pin $pin, Request $req, EntityManagerInterface $em): Response
     {
-        $this->checkUser();
+        if(!$this->isUserLoggedIn()){
+            return $this->redirectToRoute('app_login');
+        }
         $form = $this->createForm(PinType::class, $pin);
         $form->handleRequest($req);
         if($form->isSubmitted() && $form->isValid()){
@@ -71,7 +77,9 @@ class PinController extends AbstractController
      */
     public function delete(Pin $pin, Request $req, EntityManagerInterface $em): Response
     {
-        $this->checkUser();
+        if(!$this->isUserLoggedIn()){
+            return $this->redirectToRoute('app_login');
+        }
         if($this->isCsrfTokenValid('pin_deletion_'.$pin->getId(), $req->request->get('csrf_token'))){
             $em->remove($pin);
             $em->flush();
@@ -86,7 +94,9 @@ class PinController extends AbstractController
      */
     public function detail(Pin $pin, Request $req): Response
     {
-        $this->checkUser();
+        if(!$this->isUserLoggedIn()){
+            return $this->redirectToRoute('app_login');
+        }
         return $this->render('pin/detail.html.twig', compact('pin'));
     }
 
@@ -97,11 +107,12 @@ class PinController extends AbstractController
      *
      * @return void
      */
-    private function checkUser()
+    private function isUserLoggedIn():bool
     {
         if(!($this->getUser())){
-            $this->addFlash('error', 'It seems like you\'re not logged in yet!');
-            return $this->redirectToRoute('app_login');
+            $this->addFlash('danger', 'It seems like you\'re not logged in yet!');
+            return false;
         }
+        return true;
     }
 }
